Refactor out a base class for Regex pattern -> Token substitutions
